pyproject.toml
[project]
name = "ai-agent"
version = "0.2.0"
description = "AI Agent with LangChain, LangGraph, and OpenRouter for structured JSON template generation"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Arsen Tymchuk", email = "arsen.tymchuk@example.com" }
]

dependencies = [
    "fastapi[standard]>=0.115.0",
    "scalar-fastapi==1.6.1",
    "sse-starlette>=3.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "faiss-cpu>=1.8.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "poethepoet>=0.24.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "src/api",
    "src/agents",
    "src/infrastructure",
    "src/application",
    "src/shared",
    "src/store",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
filterwarnings = [
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = [
    "E501",  # Line too long (handled by formatting or unavoidable in prompts)
]

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = "3.12"
strict = false
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = false
disable_error_code = ["misc", "override", "assignment", "arg-type", "var-annotated", "operator", "call-arg"]

[[tool.mypy.overrides]]
module = [
    "pytest",
    "_pytest.*"
]
follow_imports = "skip"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "langchain_core.*",
    "langchain.*",
    "langchain_community.*",
    "langchain_openai.*",
    "langgraph.*",
    "faiss.*",
    "scalar_fastapi.*",
    "sse_starlette.*",
]
ignore_missing_imports = true

[tool.poe.tasks]
lint = "ruff check src tests"
format = "ruff format src tests"
typecheck = "mypy src tests"
test = "pytest"
pre-commit = "pre-commit run --all-files"
check = ["lint", "typecheck", "test"]
serve = "uvicorn src.main:app --reload"
